Reorganize src folder
The next few commits will make BioNEV installable as a package. This commit reorganizes the src/ folder to have the community standard layout, in which the name of the package is the only folder within src/

This commit also adds __init__.py files in folders where they were missing. This will be a good place to put submodule-level documentation